// frontend/src/core/api/model/changePasswordDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Dentizy API
 * API Documentation untuk Sistem Manajemen Klinik Gigi
 * OpenAPI spec version: 1.0
 */

export interface ChangePasswordDto {
  /** Password lama yang saat ini digunakan user */
  oldPassword: string;
  /**
   * Password baru yang akan digunakan. Minimal 8 karakter, harus mengandung huruf besar, huruf kecil, angka, dan karakter spesial
   * @minLength 8
   */
  newPassword: string;
  /** Konfirmasi password baru, harus sama dengan field newPassword */
  confirmPassword: string;
}