// src/core/api/model/updateProfileDto.ts
/**
 * Generated by orval v8.0.0 🍺
 * Do not edit manually.
 * Dentizy API
 * API Documentation untuk Sistem Manajemen Klinik Gigi
 * OpenAPI spec version: 1.0
 */

export interface UpdateProfileDto {
  /**
   * Username baru pengguna
   * @minLength 5
   */
  username?: string;
  /**
   * Nama lengkap baru pengguna
   * @minLength 3
   */
  nama_lengkap?: string;
  /** URL foto profil (dari response upload) */
  profile_photo?: string;
}